BMS/requirements.txt
asn1crypto==1.3.0
astroid==2.3.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==2.2.5
django-cors-headers==3.2.0
django-redis==4.11.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-extensions==0.6.0
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
openapi-codec==1.3.2
Pillow==6.2.1
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
PySocks==1.7.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
